// admin/src/app/gohappygobackend/models/update-platform-pricing-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdatePlatformPricingDto {

  /**
   * Platform fee in EUR (not percentage)
   */
  fee?: number;

  /**
   * Lower bound of the price range in EUR
   */
  lowerBound?: number;

  /**
   * Upper bound of the price range in EUR
   */
  upperBound?: number;
}