// admin/src/app/gohappygobackend/models/create-platform-pricing-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreatePlatformPricingDto {

  /**
   * Platform fee in EUR (not percentage)
   */
  fee: number;

  /**
   * Lower bound of the price range in EUR
   */
  lowerBound: number;

  /**
   * Upper bound of the price range in EUR
   */
  upperBound: number;
}